Shows state for Kotlin Components
Summary:
Shows state for Kotlin Components.

Since state for Kotlin Components not named, for now the Flipper UI debugger will show them keyed by their index.

StateContainer for a KComponent will now be written to the scoped info as well. This has to be different from SpecGen Components because they create a state container before render/create-initial-state is invoked; where as KComponents create state during render. So, the state container is written to inside `useState`.

Perhaps eventually, a compiler plugin could be used to pick the name of the val initialised by `useState`, and encode that for the corresponding index in KStateContainer of that component (for debug builds).

Support for common composite value types will be added in upcoming diffs.

Reviewed By: mihaelao

Differential Revision: D39473656

fbshipit-source-id: 56c0acef7aa9211a16ac2ca5e4b1de636cb30eca