Make Post to Choreographer the default batching strategy
Summary:
After our experiment we are deciding to go with the Choreographer approach as our batching strategy and to be the default for the Litho framework.

In this diff we:
- Remove the other approaches.
- Remove the test configurations from the Litho Test Runner.
- Update `KBatchedStateUpdatesTest` to work only for the batched approach.
- Make the `PostToChoreographer` approach as default.

Note that we didn't remove totally the possibility of easily fallbacking to no-batching. I will leave this for some weeks in case we need to experiment a bit more, or give the ability of no batching to some specific surface.

Reviewed By: adityasharat

Differential Revision: D39535068

fbshipit-source-id: c606a57d0379faf9b25c433fe69c5907600a0d6e