Sets the root host in layout before unmount all
Summary:
Sets the root host in layout before unmount all.

Unmount all will remove views from the hierarchy, and this WILL request layout; so it is possible that this layout request can trigger another mount pass on the MountState while it is already unmounting content from it. When control returns back to unmount all it could lead to crashes because Mount Item being unmounted will not be present in HostView.

This diff marks the root host in layout which stops the layout request from bubbling up!

Reviewed By: nicous

Differential Revision: D39771350

fbshipit-source-id: 8a2c9ac63acfa6a4247a65065220058d21d18b06