Fix @OnRegisterRanges not working after Kotlin conversion
Summary:
# Symptom
 `OnRegisterRanges` not working after some components were converted from Java to Kotlin.

# Issue
The issue here is that the code relies on the name of the method which is now appended with the package name. This is a fragile comparison since the component is free to use any method names for these lifecycle methods as long as the correct annotation is applied.

# Fix
Use the annotations to determine the lifecycle methods.

Reviewed By: adityasharat

Differential Revision: D39863514

fbshipit-source-id: 52e5ea128ffb4e4640ea98bf724156da19d96539