Ensure calculation-state-context is cleared after tests run
Summary:
Added an "afterRunnable" list to MockitoWithAutoMockSupport tests that trigger after each test. Now ensuring when a ComponentContext is initialised for a test, a runnable is queued to clear the state-context after the test runs.

This ensures we don't experience memory leaks during tests when the test completes but the state-context isn't cleared.

Reviewed By: adityasharat

Differential Revision: D39542167

fbshipit-source-id: a3bf929cbab1fb2a4279fbb82356cf2833d7d7e9