Use ThreadLocal in ComponentTree to hold RSC & LSC, and delegate to that from ComponentContext
Summary:
After investigation, it was determined that it is no longer viable to store RenderStateContext (RSC) and LayoutStateContext (LSC) on the ComponentContext (CC) moving forward with splitting render and layout.

The main reason for this is that scoped CCs can be stored on LithoNodes that are then reused by subsequent layouts. Since when calculating these layout we won't have direct access to these CCs, there's no way of storing critical layout-calculation-specific information there (i.e., RSC & LSC) when reused.

Another challenge is that a single ComponentTree (CT) can be calculating multiple layouts simultaneously in multiple different threads - so the alternative to storing the RSC & LSC must be specific to a CT and a thread.

This diff introduces a change whereby the RSC & LSC are held in a ThreadLocal in CT via a shared interface.

All existing CC methods for getting the RSC / LSC or derived information now delegate to the methods in CT that fetch it and cast to the correct type from the ThreadLocal.

** CT will check if there's already an existing instance before setting a new one prior to layout calculation. If there was an existing one, then rather than remove the reference, the previous instance will be restored.
This can happen when, for example, rendering triggers a state-update which in turn triggers another layout calculation - and is only relevant for tests.

Reviewed By: astreet

Differential Revision: D39347187

fbshipit-source-id: 20579bd21b2c80efab74308bc20ed28105838878