Merge Render & Layout caches into one class
Summary:
In upcoming diffs, we will introduce nested ComponentContexts, RenderStateContexts & LayoutStateContexts. All of these contexts will need to be reusing the same cached results.

To achieve this, we must merge both cache classes into one class, that is always read+write (previously, layout-phase caches were read-only). This is done so that when performing render in layout phase from measureNestedTree, we still have access to the caches from the render phase, and can continue to write new caches there.

However, to avoid writing to the cache created in the render phase itself, thus potentially affecting render results, we implement this via cache delegation - the layout-phase cache is created as a unique instance capable of writing its own new caches, but when reading, we allow this cache to delegate to the render-phase cache thus gaining access to the render caches, while allowing to write new caches created in the layout phase.

As an additional safety measure, caches can be locked-for-write, throwing an exception if we attempt to write more caches into it. Lock is currently not called, as this can only work once we implement nested RSC & LSC for component.measure & measureNestedTree.

Also, a small addition - RSC now holds the layout-version int with a public getter (to be used to subsequent diffs).

Reviewed By: astreet

Differential Revision: D39271798

fbshipit-source-id: b172d8b8930553931987bfd1f94ad11250ef4e61