Implement nested operations whereby RSC is set during render (and nested render) phase, and LSC is set during layout (and nested layout) phase
Summary:
In this diff, an RSC is set on the CC during the render phase, and a LSC is set during layout phase. In between these phases, the respective context is cleared.

This diff also introduces this pattern for "nested" operations, such as during component.measure & layout.measureNestedTree - which both save the "current" state-context, then set their own sequence of RSC & LSC, and finish by restoring the previous state-context.

For tests, the major difference here is that we no longer simply need to set a LSC-for-tests. Instead, an RSC is usually expected. To that end, a new set-RSC-for-tests method has been added to CC.

All litho tests and test helpers have been adapted to this new paradigm

Reviewed By: astreet

Differential Revision: D39502953

fbshipit-source-id: 7630fb996ef09d46096dcdb317834296073f4d94