Fix tests broken by previous commit
Summary: The change in the previous change made it so the Text component would default to an empty testKey when constructed through the Kotlin constructor method. This chnage will only set the `testKey` property if it is passed in. It seems the tests are depending on some automatically generated value for testKey of Text component when it is not explicitly set.

Differential Revision: D39975691

fbshipit-source-id: bc393d5d99754ff1d46c710453c9429e66a2738d