Make the LithoLifecycleProvider in LithoView/ComponentTree constructor param not null.
Summary:
The goal of this diff is to remove the `Nullable` annotation from the `LithoLifecycleProvider` in the `create` method for `LithoView` and `ComponentTree`.

We want to make the API more explicit in the need of passing in a LifecycleProvider, so this is the first step. On future next steps we are planning to deprecate the other constructors.

I have identified some usages that relied in the `null` parameter or that used the CT `create` method instead of the `LithoView` and fixed them.

Reviewed By: pentiumao, adityasharat

Differential Revision: D39936430

fbshipit-source-id: 9dde8ecd523982ad36dc48d55143bec20596152c