Create generic base class for LayoutStateFuture, and have CT's LayoutStateFuture extend this base class
Summary:
In the end-state, we will eventually have 2 different "futures", one for render, and one for layout.
RSC & LSC will both hold a reference to the future so that we can know if the future was released, interrupted, etc.

As a 1st step towards this end-state, we must have a way to A/B test the new futures with the existing LayoutStateFuture.

To that end, this diff introduces a generic base-class for this future - "ComponentTreeFuture". This is a generic class which accepts a generic type that has the potential of being partial (PotentialPartialResult).

LayoutStateFuture has been adapted to now extend this new class.

No behaviour changes are expected at this stage.

Reviewed By: adityasharat

Differential Revision: D40020026

fbshipit-source-id: 7e304d5d508b1181031f9950cf0cc058f4d35835