Add isEquivalentTo to base-class API, and implement in RTF & LTF
Summary:
Soon we will want the existing LSF to be interchangeable with the new futures. For this to work, we need isEquivalentTo to become a base-class method, as it's used in CT to check whether or not we can "swap out" futures.

This diff adds this method as an abstract method in the base class, and adds implementation to all 3 sub-classes.

Reviewed By: adityasharat

Differential Revision: D40145183

fbshipit-source-id: 02e67ea11cc16238f1a2fc0f6391c2db74036e25