Support DebugComponent toString
Summary:
## This Diff
In order to do the printing, I need to transform a collection item (component) into a debug component and then print the hierarchy.  Most of this code is reusable from the existing one, if we expose another function that takes the root input.

In order to do nice formatting, I currently add additional depth to the hierarchy, so I need to expose setting the depth.

## This Stack
Adds support for printing a recycler view hierarchy in the collection testing API.  For more details, see next diff.

Reviewed By: fabiocarballo

Differential Revision: D40286205

fbshipit-source-id: 78ad109486baea2637c4569c16aec404a7303410