Add screen coverage to VisibilityChangedEvent
Summary:
# Context
Viewport coverage signals are needed to test the IG VPV definition on Most Recent Tab. This series of diffs will add these signals to Merlin Manager for server-side processing.

Project doc for further context:
https://docs.google.com/document/d/1d3jCPI34FsxbJCDH-7Nd_DasbbHV1KTdZ6bRrLvtxiQ/edit?usp=sharing

# This diff

This diff adds the signals for host view width and height in VisibilityChangedEvent. The plan is to use this event for detecting changes in visibility and sending messages to LogDB when visibility goes over or under 50%, as well as inside or outside of 100% viewport coverage.

Reviewed By: adityasharat

Differential Revision: D40437615

fbshipit-source-id: bb7d1e8f5a2406b585293b648481f8c13be7d02e