Convert TransparencyEnabledCardClipSpec from MountSpec to MountableComponent
Summary:
This diff adds a new mountable component (`ExperimentalTransparencyEnabledCardClip`) equivalent to `TransparencyEnabledCardClipSpec`.

NOTE: As it is possible to notice from the video attached below, it looks like there is an issue with the handling of the clipping color when all the corners are rounded. The behaviour is consistent with both component as it is probably inherent to **TransparencyEnabledCardClipDrawable** which is used by both components. A follow up diff will be created to address the problem.

Differential Revision: D40550460

fbshipit-source-id: 46d97e7a001390b02fc71391441e56d858d9dde8