Allow TransparencyEnabledCardClipDrawable to take clipping color into account
Summary:
The **TransparencyEnabledCardClipDrawable** class has a different implementation of the `draw()` method when all the 4 disable clip properties are set to false.
The implementation uses the `drawRoundRect()` APIs conditionally dependending on the Android version it is currently running on (implemented in this diff D13438702 (https://github.com/facebook/litho/commit/35521a4a2669b5435e5f5103efeac4be467ba7f8)) but it did not consider the clipping color. In that case it was only drawing the clipped background rect leaving the corners transparent.

In this diff we firstly draw a rect that will be used for the corners before applying the background on top.

Differential Revision: D40553971

fbshipit-source-id: 1b36e0d9d86937fb0d748f4105b82710d5299fa0