Convert HorizontalScrollSpec from MountSpec to MountableComponent
Summary:
This diff adds a new mountable component (`ExperimentalHorizontalScroll`) equivalent to `HorizontalScrollSpec`.

NOTE: The HorizontalScrollSpec had an inner class for the HorizontalScrollView and in this diff it has been kept internal. A follow up diff will move the View to its own class to avoid references of the MountSpec class into the new Mountable. There was also an issue that was identified during this migration regarding the layout direction not being accessible anymore in the measure() method. A follow up diff will be created to address the issue.

Differential Revision: D40630415

fbshipit-source-id: 07c6ad0ea9e5c7d4880a4753f5847897f528a0e8