Allow measure() to get access to LayoutDirection property
Summary:
In this diff we extend the LayoutContext object to allow passing additional layout information to the MountableComponent measuring function.

While working on the HorizontalScroll migration we realised there were some properties missing during the measuring phase that are available for MountSpecs but not for MountableComponents.

We added an ExtraBundle to the RenderState.LayoutContext where we can store such properties.

Differential Revision: D40639859

fbshipit-source-id: 6fb061dc7634d073608c75f176131e65a5521c75