Convert HorizontalScrollSpec from MountSpec to MountableComponent
Summary:
With this diff we move the `HorizontalScrollLithoView` class to its own file so that we can remove references of `HorizontalScrollSpec` from `ExperimentalHorizontalScroll` MountableComponent.

Together with the View class, also the inner `ScrollPosition` and `OnScrollChangeListener` classes have been moved in this diff.
The listener had not actived usages so far impacting the rest of the code, while the other two classes were only used internally by the components.

Differential Revision: D40633088

fbshipit-source-id: 387c8f58e871cc5e903e3ed44e06e3da57c8ea2a