Remove transition field from StateContainer
Summary:
This is necessary to make StateContainers immutable since otherwise we need to transfer the transition in addition to state in transferState.

The approach is not super nice, but helps isolate the API as much as possible: we need to return a Transition from applyStateUpdate so TransitionContainer provides a new applyStateUpdateWithTransition that should be used instead of applyStateUpdate. This, for example, prevents Transition from leaking to the Sections/Surface APIs.

Differential Revision: D40639078

fbshipit-source-id: 0a18914513a2da54377c2c0672f6fefc9f52eab4