Begin implementation for running LayoutTreeFuture after resolve completes.
Summary:
Beginning work on running the LayoutTreeFuture. This diff adds logic for running the LTF either sync or async, collects the fields needed async and then starts the sequence.

The sequence is defined as running the future, performing validations on the resulting LayoutState, setting the output size, and committing the LayoutState.

Running the future and committing the LS are still unimplemented.

For more information, see here: https://fburl.com/gdoc/afx72arw

Reviewed By: adityasharat

Differential Revision: D40341593

fbshipit-source-id: c98efbb5dd621f194e045a7b96ef10da348c8b36