Implement logic for committing LithoResolutionResult
Summary:
After the LithoResolutionResult is created, it is committed in the ComponentTree. This is only done if the current committed one is either null, or has a lower resolveVersion than the new one.

Added resolve-version to LithoResolutionResult, and refactored naming in RenderTreeFuture to Resolve-Version (rather than Layout-Version)

For more details, see here: https://fburl.com/gdoc/g13t1ef2

Reviewed By: adityasharat

Differential Revision: D40341002

fbshipit-source-id: e02637cef0238098d432cedd5b2ba8ac269c099d