Implement logic to skip resolve if triggered by measure / setSizeSpec and a resolution-result is already available.
Summary:
If a resolution-result is already available, and a calculation request came in with source = measure or setSizeSpec (sync or async), we can effectively skip the resolve step and move directly to layout.

In addition, if a measure is requested but no resolution-result is available (and none are being calculated), we stop and trigger a resolution instead.

For more details, see here: https://fburl.com/gdoc/bdm4yy93

Reviewed By: adityasharat

Differential Revision: D40346210

fbshipit-source-id: ded1e41122dcd566f9369dcfc24e7a5bd263fccb