Implement logic for running LayoutTreeFuture
Summary:
Logic for creating and running a LayoutTreeFuture is added in this diff. For async runs, we abort if there's already a running LTF with the same resolution-result, width and height.

A few minor refactors were added to ensure the LTF is getting all the fields it needs, and doesn't receive duplicate fields that are already provided via the resolution result.

For more details, see here: https://fburl.com/gdoc/au5eh3zu

Reviewed By: adityasharat

Differential Revision: D40345239

fbshipit-source-id: f295a6fbe2b995bae2502d23e6398fd55a1875d1