Use StateContainers from StateHandler as source-of-truth StateContainers for specs
Summary:
Instead of creating a StateContainer in ScopedComponentInfo and transfering state to it, we now just return the StateContainer from the StateHandler which will already have the latest state (once we ship applyStateUpdatesEarly).

The most interesting part of this diff is we've had to replace the 'mark state for global key needed' logic which was in applyStateUpdates previously. It's moved to useState (for KComponents) and to createScopedContext (for Specs). Open to something cleaner as I don't love this, might try to think of something better tomorrow.

Reviewed By: adityasharat

Differential Revision: D40031440

fbshipit-source-id: e7d9b9414ef5904e294a06853a76f21bd88c2e5b