Update EventHandler rebinding logic to just rebind EventDispatchInfo for Sections
Summary:
This diff updates EventHandler re-binding for EventHandlers created from Sections. It's a pretty complete reimplementation of the current logic.

The basic idea here is:
1. During a changeset calculation, we create Spec-generated EventHandlers with the latest DispatchInfo (ComponentContext and EventDispatcher) and record all these EventHandlers in a list local to this computation
2. Canonicalization: If we commit this calculation, we update the collected EventHandler's DispatchInfos to be the 'canonical' DispatchInfos we already have for that global key (if we don't have one, that DispatchInfo becomes the canonical DispatchInfo for that key)
3. Re-bind: Finally, we update the canonical DispatchInfos for each global key to the latest ComponentContext/EventDispatcher.

By having all committed EventHandlers for a global key reference the same DispatchInfo, we can update them all at the same time when a new changeset is committed.

EventHandlersController2 is temporary and will be deleted when we land the same integration for components.

The difference from D38466417 is that this diff properly handles the case when a changeset calculation creates an EventHandler and then invokes it during that calculation (e.g. RenderEvents) - this is why we first create EventHandlers with a new DispatchInfo and then 'canonicalize' it after commit.

Reviewed By: adityasharat

Differential Revision: D40735730

fbshipit-source-id: 1c124fadde77ea5fea869b3f946e1c983b6ac781