When a sync LayoutTreeFuture is equivalent to a running one, hang on the running one until it completes.
Summary:
This behaviour mimics the existing LayoutStateFutures. Even tho we can't interrupt LayoutTreeFutures, we can still force the current running one to have higher priority and then hang on it until it finishes.

Raising the thread priority is already done in the base class implementation. This diff simply ensures if an equivalent LTF is found, it is reused, and not discarded until the wait-count on it reaches zero.

Also, removed now unused isSync field from LTF

Reviewed By: astreet

Differential Revision: D40591016

fbshipit-source-id: 27e2f0a0fdce78c0b0b430700749a2dd47b10747