Separate logic for using split thread handlers for resolve and layout into a separate config
Summary:
To allow us to test split thread-handlers separately, this diff introduces a new config specifically for using split thread handlers.

When disabled, any access to the resolve-thread-handler (or mutex) will go thru a method that will provide either the resolve one if the config is enabled, or the layout one if its disabled.

Few tests have been adjusted to get this to work - specifically, tests that need to "runOneTask" must now run 2 tasks.

Reviewed By: adityasharat

Differential Revision: D40384141

fbshipit-source-id: 613a70490c270d9423f6b30289ae5840034c8591