Add new test for racing setSizeSpec from separate background threads
Summary:
Added new test that checks for racing setSizeSpecs with the same values. The test verifies the behaviour wherein a running LayoutTreeFuture is reused if it has the same fields (width / height) as a new one that is being triggered.

The test achieves this logic by ensuring that the 2nd setSizeSpec isn't called until the 1st one is already being processed.

To achieve this, the test-spec introduced in the previous diff was elaborated to include listeners for onPrepare and onMeasure, where semaphores can be added to force exact timings.

Reviewed By: astreet

Differential Revision: D40635695

fbshipit-source-id: 8da672c52d039a754a8ff251e86b2c950eb81a6d