Add new test for racing async and sync setSizeSpecs
Summary:
When an async layout is requested while an equivalent sync layout is in progress, ComponentTree will ignore the async request completely.

This diff adds a test that introduces such a scenario, triggering a sync layout from a bg thread, and queuing an async layout with the same size specs.

The test ensures the timing by forcing the async tasks to wait until the sync measure starts, and then forcing the sync measure to wait until the async tasks have finished draining.

We then verify that only 1 measure happened, and with the expected values.

Reviewed By: astreet

Differential Revision: D40675985

fbshipit-source-id: a38508c5203c4a16735f5a0ea93cad9d9769825e