Add 2 new tests to ensure async render and layout (respectively) are promoted to the main thread
Summary:
This diff adds two similar tests that ensure that when a sync request comes in while an equivalent async request is running during render (test 1) or layout (test 2), the process is promoted to the UI thread rather than executing twice.

While writing these tests, a flaw in the split futures logic was found wherein a render would only be skipped if the source was from a measure-only call, meaning calling setRootAndSizeSpec (sync) while an equivalent async process is calculating the layout, another render operation would happen since it's not a measure-only source.

To resolve this, the logic was adapted in this diff to disregard the calculation source, and instead only check if the root is equivalent. In addition, we also check that the tree-props are the same.

Reviewed By: adityasharat

Differential Revision: D40714479

fbshipit-source-id: d9f957b142ebd551512d138651fd0604af7779e0