Add RETAIN_MAXIMUM_RANGE RecyclingStrategy
Summary:
This is my first attempt at improving thrash in creating and destroying views while scrolling in RecyclerBinder when it has a dynamic range size.

Here, my strategy is to allow the dynamic range size logic to continue as-is. This creates a range size based on the number of items currently on screen, which can vary on news feed based on feed unit type. Range sizes seem to quickly fluctuate back and forth sometimes - usually between 7 and 14 depending on how many items are visible.

Rather than trying to slow down the pace of increase or decrease of the range size in this strategy, I'm instead limiting the removal of out of range items after a range descrease.

In the default strategy, every scroll or visible range change triggers a range size calculation, and a call to computeRange that iterates over all items in the adapter and:

  if the index of the item is in our new range OR is sticky OR is visible:
    if the layout of the item is not already valid:
      calculate layout
  else
    release the component tree of item if it's valid

That strategy creates situations where this can happen:

* 3 items are on screen so we use a larger range size, say index +/- 14
* We create layouts for all items in our range
* We scroll a little, so only 2 items are on screen and we reduce our range size to index +/- 7
* We destroy all the component trees / layouts we just calculated of index +/- 7-14
* We scroll and either increase range size again, creating the same layouts again, or we scroll to hit index +8 and create the same layout we already created and destroyed.

In this new RETAIN_MAXIMUM_RANGE strategy, I pass a flag to a fork of computeRangeLayoutAt to tell it whether it's allowed to delete out of range items. The idea here is to handle abrupt range increases by retaining layouts calculated in those larger ranges and only delete them when we move outside that larger range.

So, if we were at index 10 and our range size jumped to 14, then we calculated layouts for indices 3-17. If our range size decreases to 7 because we scroll a little but we're still at index 10, then instead of destroying layouts for indices (approx) 3, 4, 5, 14, 15, 16, 17 - we just keep them and skip deleting them.

Once we scroll beyond the range of 3-17 in either direction, then we allow deleting out of range items.

To achieve this, I keep track of the lowest seen rangeStart value and the paired highest seen rangeEnd value on each call to computeRange. I only allow deletions during a subsequent traversal iff the rangeStart is < lowest seen rangeStart OR rangeEnd > highest seen rangeEnd. In that case, I reset highest/lowest seen range values to the rangeStart/End passed in to that traversal.

I suppress deletions by passing a boolean flag to a fork of computeRangeLayoutAt. I forked it because of this change and 2 other optimizations:

1) I skip calculating sizes for componentHolders that are out of range, since that info is unused in that case. In local testing, I saw these calls take about 15000ns per index
2) since off-ui-thread calls to computeLayout result in each index deletion being posted to the UI thread, I refactored out the validity checks that weren't marked UiThread and call them prior to posting to UI thread, to save on those calls.

Reviewed By: fabiocarballo

Differential Revision: D40737977

fbshipit-source-id: d2f77629b835541cd65ca41029fbe4da69d4b75b