When setRoot is called sync, but no size-specs are set on the CT, covert the operation to async
Summary:
In "control" behaviour, setting root without any size-specs set would do nothing, but in "test" behaviour, this will trigger render. To avoid regressing "test" behaviour, we add an optimisation in this case wherein the operation is forced to the background, thus avoiding blocking the main thread.

This diff introduces this change, and adjusts existing tests that assumed this operation would trigger render sync by simply running to end of tasks where possible, or adding a setSizeSpec prior to setRoot where not.

Reviewed By: astreet

Differential Revision: D40715472

fbshipit-source-id: 86d6abec7e90ad3479c4c7237d522c20dd9679ca