Add test that verifies render is skipped when tree-props are set and unchanged
Summary:
The logic to skip render includes a check to verify that the tree-props are the same. This check is done by ref, so if for instance, the CT logic changes to create a copy of the tree-props, this logic will fail.

This diff introduces a test that verifies that render is properly skipped when setting the same root again with unchanged tree props.

To verify this works, the code in CT was edited locally to make a copy of TreeProps before starting the render process, and as expected, this test failed.

Reviewed By: astreet

Differential Revision: D40757901

fbshipit-source-id: 632d9f03cdefd312db17c86551327ff88d8ab8ad