Adapt tests to no longer use shortWait
Summary:
Now using new mechanism of future-execution-listener to avoid the need for having "shortWaits" in several tests. Rather than adding a short-wait to ensure the code reaches a point where a future was reused, we can now explicitly add latches when this happens.

shortWait method was removed.

Also - since "FutureType" enum was not needed, it was also removed from the new API.

Reviewed By: adityasharat

Differential Revision: D40851253

fbshipit-source-id: c3e2c04314a04180b2a361c45d825154c31d9b84