Always use provided width / height spec when performing sync operation
Summary:
Calling setRootAndSizeSpecSync must always use the provided width / height spec for the calculation, especially for cases with nested component-trees. This diff ensures a provided width / size spec is always used when reaching the layout operation.

By doing this, we ensure race-conditions of 2 sync operations will always result in layout using the specs that the render operation was triggered with (rather than grabbing whatever was last set on the ComponentTree)

Reviewed By: adityasharat

Differential Revision: D40855095

fbshipit-source-id: 6e7f40b5b99e554b9c83280e230c72efec74e920