Refactor common sequences for tracking and running TreeFutures for resolve and layout
Summary:
The logic for tracking and running RenderTreeFutures and LayoutTreeFutures is essentially identical. With that in mind, this diff adds a common static method that both sequences can use, rather than duplicating the code twice.

Note: Prior to this change, LayoutTreeFuture management alone included an optimisation that would skip the whole process (return null) if an equivalent future was running and the new request was async. This optimisation was tested for resolve as well, and seems to work fine. By making both processes use a shared method, we gain this optimisation for resolve as well.

Some methods have been renamed to make the flow cleared. Methods that managed the futures for each individual step have been removed in favour of shared static method.

The "control" LayoutStateFuture sequence has not been altered, despite sharing a common flow with the new method. The reasons for this are:
1. Avoid unintentional / unexpected behaviour with control, thus keeping control as pure as possible
2. Avoid optimising control behaviour (described above) so that test behaviour can be properly compared

Reviewed By: astreet

Differential Revision: D40716988

fbshipit-source-id: 00cc4f178c57a6f7fcf4eefebd49ab98b7df8b46