Enable accessibility events to be triggered for MountableComponents
Summary:
While  working on adding a missing accessibility event for MountableComponents I was unable to receive events I specified on the component through its style.

The current logic sets the AccessibilityDelegate on the ComponentHost but it will not have children until they are actually mounted and the NodeInfo have been added as tags.

In this diff we add the AccessibilityViewDelegate to Views at mount time if accessibility is enabled globally.

NOTE: Views are tagged with a specific id where the NodeInfo (accessibility info) are stored and their delegate is cleared whenever they are unmounted. This logic lives in the `LithoViewAttributesExtension` class. The ComponentHost is responsible of removing its own AccessibilityDelegate. In this diff we also moved the node tag to the ComponentHost class to make it easier to reference.

Reviewed By: adityasharat

Differential Revision: D40757764

fbshipit-source-id: 3532bd31bdd4979c913d91861726b958c8fb43e5