Add gradle configuration to generate dokka docs
Summary:
Docs can be generated via `./gradlew dokkaHtmlMultiModule`. These are meant to replace our current javadocs and javadoc generation process.

Note, this adds a kotlin plugin dependency in gradle to some of our core libraries. However, this seems to be required to get the dokka plugin to run at all on java-only modules.

Reviewed By: apowolny

Differential Revision: D41033765

fbshipit-source-id: 6430d90cf6a87fb8a94ec86ea35cdc5230325dc9