Add shouldUpdate that receives newMountable only
Summary:
This is an API improvement which makes the `shouldUpdate` method signature less bug prone.

I noticed this when I made a mistake like this:

```
class MyMountable(
  val text: String
) {

   override fun shouldUpdate(currentMountable, newMountable, ...): Boolean {
      return text != newMountable.text
  }
}
```

In this case, my code was not referring to the `currentMountable` params, but to my own class params. This could potentially even have the same correct result (even though I'm not sure), but can bring some confusion given the different possibilities.

The change I suggest would make the signature of `shouldUpdate` to drop the `currentMountable` param, forcing people to use their class params:

```
class MyMountable(
  val text: String
) {

   override fun shouldUpdate(newMountable, ...): Boolean {
      return text != newMountable.text
  }
}
```

Reviewed By: adityasharat

Differential Revision: D40980982

fbshipit-source-id: 10a352e712aa4833bc7268a328d8b3e48ac20de1