Always use ComponentTree's width / height specs at initiation time for resolve and layout.
Summary:
Change logic for split futures to always use the width and height specs on the CT at initiation time (i.e., when setRoot / AndSizeSpecs is called) throughout the whole process, even in async. This change means that we're now passing the width and height spec always, not only when sync, and regardless of whether or not the output is null.

The width & height specs must now pass into the runnables for resolve and layout, and are used throughout the whole process.

Reviewed By: adityasharat

Differential Revision: D41157075

fbshipit-source-id: 5a9677e564a09e54974760ced7c1a24a6825b42c