Minor correction to some tests: running to end of bg-tasks in bg-thread, rather than also running to end of main-thread tasks.
Summary:
Using LegacyLithoViewRule's idle() method also runs to the end of main-thread tasks. This will cause an exception sometimes if this process manages to reach the point where it updates the main-thread-layout-state, by hitting an assertMainThread.

This diff adds a new method to LegacyLithoViewRule that only runs to the end of background tasks, and updates several tests to use when called from a bg-thread. These tests also ensure the bg-thread finishes before the test ends.

Also, added some minor refactors to some tests in SplitFuturesTest

Reviewed By: adityasharat

Differential Revision: D41183400

fbshipit-source-id: 1b3df5930910001bf384ab51eacc1c58560b05e6