add isTextTruncated() in RCTextView
Summary:
extending RCTextView with new method to determine whether the text is truncated.
this method check first if we explicitly marked the text as truncated. We do it when a custom truncation string is used, since a new layout is created and otherwise we have no info about its truncation later.

Differential Revision: D41069272

fbshipit-source-id: bb7f97fe3f15f4bf6fb82b6d48983f4114db1fa5