Fix to LayoutTreeFuture when returning a null LayoutState, avoid throwing when using cancellable futures.
Summary:
In ComponentWarmer flows, a prepare might be cancelled while in progress. This means that a running LayoutTreeFuture might get released before the ComponentTree is released, causing an exception.

In control flow, this is addressed by avoiding to throw an exception is use-cancellable-futures config is enabled. For now, we mimic this behaviour for test flows.

Reviewed By: adityasharat

Differential Revision: D41264270

fbshipit-source-id: 921331575d39c9b5dd8fde145616e47805549ca8