Fix two race conditions in RecyclerBinder.
If the collection has not completed initial layout by the time onNewWorkingRange() is called, it will short circuit and not be called again.

If a collection item has not completed initial layout by the time we attempt to check if it belongs to a working range, the call noop-s and is not attempted again.

In each case, save the working range and apply it when the layout completes.