Handle nested trees when extracting bounds for DebugComponent
Summary:
Litho will create a separate LithoLayoutResult for nested trees. The nested component will appear as a head component on the LithoNodes of both this LithoLayouResult as well as the nested result's.
This is handled in DebugComponent when extracting children by ignoring the head component if it's associated with a nested tree and moving the component index up (https://fburl.com/code/jkymj1od).

The bounds of the nested component are extracted from the NestedTreeHolderResult node, but any margins that it needs to account for are applied on the nested LithoLayoutResult. We need to account for that margin on the NestedTreeHolderResult when extracting the bounds.

Reviewed By: adityasharat

Differential Revision: D41339707

fbshipit-source-id: f4d7b4fd31984135428eec5d710b44c864a39a47