Prototype Printing Recycler View x DebugComponent info
Summary:
## This Diff

This diff is an RFC which prototypes adding the ability to print the recycler view hierarchy in test collection tests.  When using the API myself a month or so ago, I found that it was hard to debug failures, especially when it was dependent on the items within the hierarchy (great example: apparently the one design system's recycler view appends a -- what feels random -- line as the first child in every usage).

Therefore, this is a prototype to print out the recycler view children (just components for now) to have better error messages with the assertions.

## Options

**Option 1: Use `ViewTree.of()` + `makeTest`**

Using `ViewTree.of()` on the recycler view allows us to print the children (`ViewTree.of(collection.recyclerView).makeString { null }`).  I found, however, that that this isn't great for the following reasons:
1. A litho view is shown before each component, which makes it harder to quickly scan through the children
2. View extra information function just gives us the view, and it *also* appends at the end of the component tree.  So, that makes it a bit harder if we want to supply additional information about the collection item itself.
3. If we ever just want to print the visible children within the recycler or just the list of fully visible children, we can't use this existing method to do so.  We'd need to print each view individually in order to do so.
4. This isn't as useful when people are using mocks in children (see last child of first component) since it just prints a `LithoView` without any children (see last child of first real world example).

One pro of this is that it basically works out of the box and just prints the view tree (albeit, the extra extra indents feel a little odd, to me)

**Option 2: Roll Our Own via `TestCollectionDescriptionHelper`**

Rolling our own description helper allows us to add more information such about the item itself such as index, id, visibility state, and more about the test collection item.  This is useful, I think at least, to get the full lay of the information for the view state.

Some of the downsides, however, are that it:
1. Requires some updates to `LithoViewTestHelper.java` to support passing in a root `DebugComponent`
2. Means that we need to update things ourselves (but really the majority of this is just delegation which I don't think is so bad)

NOTE: Personally, I prefer option 2 as someone who has used the API and wished I had this output (from the 1 time I wrote some tests for product code and from the time I was adding new assertions to the actual test collection API).

## Asking for Feedback On
- If people have options preferences (`ViewTree.of` vs. `TestCollectionDescriptionHelper`)
- If they prefer `TestCollectionDescriptionHelper`, if they like it with or without the new line between collection items
- Anything else

This is a prototype, so I need to write tests, and I guess also support the view child case.

Reviewed By: fabiocarballo

Differential Revision: D39990902

fbshipit-source-id: 949f8e27b43d08b20764f9f835489902aeb3dbea